fixing pyang.bat on Windows
Major: the "%cd%" in the pyang.bat script returned the current working directory - hence the path "%cd%\pyang" pointed mostly to an non existing file. The new "%~dp0" returns the current directory location of the pyang.bat itself - that should the the same directory where the "pyang" file is installed.
Mimor: The former Windows pyang.bat script has executed a hardcoded "python" intepretter, while in a real world Windows setup might be "python", "python2" or "python3". The "sys.executable" should point to the correct name and path to the Python executable at the time of the installation.